AvatorStudio’s  AvatarAnimator and FaceController change to compile
