+++ .gitignore
@@ -92,3 +92,4 @@ Thumbs.db
 UserSettings/Layouts/default-2021.dwlt
 
 VersionControlSettings.asset
+ProjectSettings
