--- Assets/Src/Scripts/Animator/FaceController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e1281378c7bf46f5aca5e3e7be29f96
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
